requirements.txt
streamlit
opencv-python-headless
numpy
tensorflow
rembg
pillow
onnxruntime